fix: implemented hamburger menu
