+++ lib/components/dashboard/UserSection.tsx
@@ -1,18 +1,7 @@
-import { DashboardMode } from '../../resources/commonTypes';
-import Notification from './Notification';
-import { useState } from 'react';
-import HamburgerMenu from './HamburgerMenu';
-
-interface Props {
-  mode: DashboardMode;
-}
-
 /**
  * User login/logout buttons.
  */
-const UserSection: React.FC<Props> = ({ mode }) => {
-  const [openHamburger, setOpenHamburger] = useState(false);
-
+const UserSection: React.FC = () => {
   return (
     <div className="fixed z-20 w-screen h-16 p-3 px-6 select-none bg-primary">
       <div className="flex flex-row items-center justify-end w-full h-full">
@@ -23,20 +12,6 @@ const UserSection: React.FC<Props> = ({ mode }) => {
           <img src="/img/logo-darker.png" alt="logo" className="mr-3 h-9"></img>
           <div>uCredit</div>
         </div>
-        <Notification />
-        <div
-          className="p-2 space-y-2 bg-gray-100 rounded shadow h-9 w-9 mx-2 cursor-pointer"
-          onClick={() => setOpenHamburger(!openHamburger)}
-        >
-          <span className="block w-5 h-0.5 bg-black"></span>
-          <span className="block w-5 h-0.5 bg-black"></span>
-          <span className="block w-5 h-0.5 bg-black"></span>
-        </div>
-        <HamburgerMenu
-          openHamburger={openHamburger}
-          setOpenHamburger={setOpenHamburger}
-          mode={mode}
-        />
       </div>
     </div>
   );
