+++ lib/components/dashboard/degree-info/Distributions.tsx
@@ -2,12 +2,10 @@ import { XIcon } from '@heroicons/react/solid';
 import { FC, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import Select from 'react-select';
-import { Major, ReviewMode } from '../../../resources/commonTypes';
+import { Major } from '../../../resources/commonTypes';
 import { selectTotalCredits } from '../../../slices/currentPlanSlice';
 import { updateInfoPopup } from '../../../slices/popupSlice';
-import { selectReviewMode } from '../../../slices/userSlice';
 import CourseBar from './CourseBar';
-import Reviewers from './Reviewers/Reviewers';
 
 /**
  * Area in the right hand plan information that shows various elements of degree progression.
@@ -20,8 +18,7 @@ const Distributions: FC<{
 }> = ({ major, userMajors, distributionBarsJSX, changeDisplayMajor }) => {
   // Component state setup.
   const totalCredits = useSelector(selectTotalCredits);
-  const [disclaimer, setDisclaimer] = useState<boolean>(false);
-  const reviewMode = useSelector(selectReviewMode);
+  const [disclaimer, setDisclaimer] = useState<boolean>(true);
   const dispatch = useDispatch();
 
   const majorOptions = userMajors.map((m, index) => ({
@@ -38,14 +35,8 @@ const Distributions: FC<{
       <div className="flex flex-row mb-3 w-full">
         <div className="self-start text-2xl font-medium">Degree Progress</div>
         {/* Degree Progress */}
-        <button
-          className="w-24 h-6 mt-1 ml-1 text-center bg-red-100 rounded"
-          onClick={() => setDisclaimer(!disclaimer)}
-        >
-          Please read
-        </button>
         <div
-          className="h-6 w-6 m-auto cursor-pointer"
+          className="h-6 w-6 m-auto mr-0 cursor-pointer"
           onClick={() => dispatch(updateInfoPopup(false))}
         >
           <XIcon />
@@ -62,33 +53,6 @@ const Distributions: FC<{
           className="z-50 w-full"
           hideSelectedOptions
         />
-      )}
-      {disclaimer && (
-        <div>
-          <b> This feature is still being refined. </b> Degree criteria on
-          uCredit is currently implemented by hand to match as closely to
-          university requirements as possible. However, there may be some
-          inconsistencies. Please use the
-          <a
-            href={getHref()}
-            className="mx-1 text-blue-400"
-            target="_blank"
-            rel="noreferrer"
-          >
-            official undergraduate advising manual
-          </a>
-          and the{' '}
-          <a
-            href="https://sis.jhu.edu/sswf/"
-            className="text-blue-400"
-            target="_blank"
-            rel="noreferrer"
-          >
-            SIS degree audit
-          </a>{' '}
-          to double check that your degree is being correctly tracked. Please
-          report any issues in the feedback form.
-        </div>
       )}
       <CourseBar
         distribution={{
@@ -109,8 +73,67 @@ const Distributions: FC<{
         bgcolor=""
       />{' '}
       {distributionBarsJSX}
-      {/* M notes: distributionsBarsJSX is where the bars except total credits are created  */}
-      {reviewMode !== ReviewMode.View && <Reviewers />}
+      {disclaimer && (
+        <div
+          id="dropdown-cta"
+          className="p-4 mt-6 bg-blue-50 rounded-lg"
+          role="alert"
+        >
+          <div className="flex items-center mb-3">
+            <span className="bg-orange-100 text-orange-800 text-sm font-semibold mr-2 px-2.5 py-0.5 rounded">
+              Beta
+            </span>
+            <button
+              type="button"
+              className="ml-auto -mx-1.5 -my-1.5 bg-blue-50 text-blue-900 rounded-lg focus:ring-2 focus:ring-blue-400 p-1 hover:bg-blue-200 inline-flex h-6 w-6"
+              data-collapse-toggle="dropdown-cta"
+              aria-label="Close"
+              onClick={() => setDisclaimer(false)}
+            >
+              <span className="sr-only">Close</span>
+              <svg
+                className="w-4 h-4"
+                fill="currentColor"
+                viewBox="0 0 20 20"
+                xmlns="http://www.w3.org/2000/svg"
+              >
+                <path
+                  fill-rule="evenodd"
+                  d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
+                  clip-rule="evenodd"
+                ></path>
+              </svg>
+            </button>
+          </div>
+          <div>
+            <p>
+              This feature is still being refined. Please report any issues in
+              the feedback form.
+            </p>
+          </div>
+          <div className="text-sm mt-4">
+            Please double check with the{' '}
+            <a
+              href={getHref()}
+              className="text-blue-900 underline hover:text-blue-800"
+              target="_blank"
+              rel="noreferrer"
+            >
+              official advising manual
+            </a>{' '}
+            and the{' '}
+            <a
+              href="https://sis.jhu.edu/sswf/"
+              className="text-blue-900 underline hover:text-blue-800"
+              target="_blank"
+              rel="noreferrer"
+            >
+              SIS degree audit
+            </a>
+            .
+          </div>
+        </div>
+      )}
     </div>
   );
 };
