+++ lib/components/dashboard/degree-info/ShareLinksPopup.tsx
@@ -21,11 +21,8 @@ const ShareLinksPopup: FC<{
 
   return (
     <>
-      <div
-        className="absolute top-0 z-30 w-min h-screen"
-        onClick={() => setURL()}
-      />
-      <div className="absolute z-40 flex items-center h-6 overflow-auto transform -translate-x-full -translate-y-1/2 rounded-md rounded-t-lg w-80 bg-gray-50">
+      <div className="z-30 w-min" onClick={() => setURL()} />
+      <div className="z-40 flex items-center h-6 overflow-auto transform absolute right-0 rounded-md rounded-t-lg w-80 bg-white">
         <div className="flex-grow flex-shrink">
           <form>
             <input
