+++ lib/components/dashboard/index.tsx
@@ -34,7 +34,6 @@ import DeleteYearPopup from '../popups/DeleteYearPopup';
 import PlanAdd from '../popups/PlanAdd';
 import CourseList from './course-list/horizontal/CourseList';
 import InfoMenu from './InfoMenu';
-import ActionBar from './degree-info/ActionBar';
 import { selectLoginCheck, selectUser } from '../../slices/userSlice';
 import axios from 'axios';
 import { getAPI } from './../../resources/assets';
@@ -45,6 +44,8 @@ import HandlePlanShareDummy from './HandlePlanShareDummy';
 import HandleUserInfoSetupDummy from './HandleUserInfoSetupDummy';
 import { DashboardMode, Plan, ReviewMode } from '../../resources/commonTypes';
 import { userService } from '../../services';
+import HamburgerMenu from './HamburgerMenu';
+import Notification from './Notification';
 
 interface Props {
   plan: Plan;
@@ -76,6 +77,7 @@ const Dashboard: React.FC<Props> = ({ plan, mode }) => {
   const [showNotif, setShowNotif] = useState<boolean>(true);
   const [formPopup, setFormPopup] = useState<boolean>(false);
   const [showHeader, setShowHeader] = useState<boolean>(true);
+  const [openHamburger, setOpenHamburger] = useState(false);
   // const [experimentPopup] = useState<boolean>(false);
   // const [displayedNumber, setDisplayedNumber] = useState<number>(3);
   // const [crement, setCrement] = useState<number>(0);
@@ -173,12 +175,11 @@ const Dashboard: React.FC<Props> = ({ plan, mode }) => {
               notifHandler={setShowNotif}
             />
           )}
-          {showHeader && <UserSection mode={DashboardMode.Planning} />}
+          {showHeader && <UserSection />}
           <div className="flex-grow w-full">
             <div className="flex flex-col w-full">
               <div className="flex flex-row thin:flex-wrap-reverse mt-[5rem] w-full h-full">
                 <div className="flex flex-col w-full">
-                  {plan ? null : <ActionBar />}
                   <div className="px-[100px]">
                     <CourseList plan={plan} mode={mode} />
                   </div>
@@ -201,6 +202,22 @@ const Dashboard: React.FC<Props> = ({ plan, mode }) => {
       <GenerateNewPlan />
       <HandleUserInfoSetupDummy plan={plan} />
       <HandlePlanShareDummy />
+      <div
+        className="z-40 p-2 space-y-2 bg-gray-100 rounded shadow h-9 w-9 mx-2 cursor-pointer absolute top-3 right-7"
+        onClick={() => setOpenHamburger(!openHamburger)}
+      >
+        <span className="block w-5 h-0.5 bg-black"></span>
+        <span className="block w-5 h-0.5 bg-black"></span>
+        <span className="block w-5 h-0.5 bg-black"></span>
+      </div>
+
+      <Notification />
+
+      <HamburgerMenu
+        openHamburger={openHamburger}
+        setOpenHamburger={setOpenHamburger}
+        mode={DashboardMode.Planning}
+      />
     </>
   );
 };
